Intro to Linear models full tutorial
I've formatted Laura's .Rmd of the tutorial into .md to be the website page. This isn't a file upload because the .Rmd to .md script we have on the website didn't work for converting it. Not sure what the problem is but it was just easier to manually edit for the time being. As it's copy pasted, formatting commits are likely to follow after this.